Update README.md
Looking for 0 interest first loan apps that are 100% legit and fast to approve in the Philippines? Whether you're a seasoned trader or a regular Pinoy needing emergency cash, you're in the right place. This 2025 guide gives you the top 10 loan apps offering 0% interest on your first loan – legal, secure, and fast.